requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.0.0
pydantic==2.12.5
email-validator==2.3.0
pandas==3.0.1
matplotlib==3.10.8
